:recycle: Refactoring and little bugfix
