+++ pincer/utils/insertion.py
@@ -5,7 +5,7 @@ from inspect import getfullargspec, Parameter, Signature
 from typing import Any, Union, Callable, Mapping, List
 
 from .types import Coro
-from ..objects.messagecontext import MessageContext
+from ..objects.message_context import MessageContext
 
 
 def should_pass_cls(call: Union[Coro, Callable[..., Any]]) -> bool:
